Update Character stats
Added first iteration of characters stats consisted of generic early FF
games. There are no usages for them right now. Moved most stats to
floats but kept them displayed as ints. Replaced attack with physical
damage

+++ Assets/Code/Combat/Calculations/DamageCalculator.cs
@@ -2,7 +2,7 @@
 {
     public static class DamageCalculator
     {
-        public static int Damage(ICharacter source, Character target, int damage, DamageType damageType)
+        public static float Damage(ICharacter source, Character target, float damage, DamageType damageType)
         {
             return damage;
         }
